server/requirements.txt
alembic==1.5.8
appdirs==1.4.4
APScheduler==3.7.0
bcrypt==3.2.0
bidict==0.21.2
boto3==1.17.76
botocore==1.20.76
cachelib==0.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
croniter==1.0.13
cryptography==3.4.7
distlib==0.3.1
dnspython==1.16.0
email-validator==1.1.2
eventlet==0.31.0
filelock==3.0.12
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-JWT==0.3.2
Flask-JWT-Extended==4.2.3
Flask-Login==0.5.0
Flask-Migrate==2.7.0
Flask-Script==2.0.6
Flask-Session==0.4.0
Flask-SocketIO==5.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
gevent==21.1.2
greenlet==1.0.0
gunicorn==20.1.0
http-ece==1.1.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
Mako==1.1.4
MarkupSafe==1.1.1
py-vapid==1.8.2
pycparser==2.20
pyjsparser==2.7.1
PyJWT==2.1.0
pypiwin32==223
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
python-engineio==4.2.0
python-socketio==5.3.0
pytz==2021.1
pywebpush==1.13.0
pywin32==300
redis==3.5.3
requests==2.25.1
rq==1.8.1
rq-scheduler==0.11.0
s3transfer==0.4.2
schedule==1.1.0
six==1.15.0
SQLAlchemy==1.4.3
tzlocal==2.1
urllib3==1.26.4
virtualenv==20.4.4
Werkzeug==1.0.1
win10toast==0.9
WTForms==2.3.3
zope.event==4.5.0
zope.interface==5.4.0
